feat: add warning for missing checks for better debugging experience (#610)
When the subproject in the configuration file contains no paths or no checks, the current behavior is that the check will fail silently as indicated in #593. This behavior is confusing for developers since the impression will be that nothing went wrong. We want to call out this issue. However, it is valid for someone to leave the checks empty for now and add more things later as the project evolves, so the behavior chosen is to add warnings at the end of the detail status to let the users know what happened.